+++ src/account/index.ts
@@ -3,6 +3,6 @@ export * from "./Account";
 export * from "./SingleKeyAccount";
 export * from "./EphemeralKeyPair";
 export * from "./KeylessAccount";
-export * from "./KeylessAccountCommon";
+export * from "./AbstractKeylessAccount";
 export * from "./FederatedKeylessAccount";
 export * from "./MultiKeyAccount";
